Code quality pass: extract settings views, remove dead code, clean up imports
- Extract 5 settings sections from UsagePopoverView (909→606 lines) into Views/Settings/
- Delete SandboxAccessManager (dead code behind never-defined APP_SANDBOX flag)
- Remove ObservableObject/Combine from NetworkMonitor (only read directly), add deinit cancel
- Remove redundant .receive(on: DispatchQueue.main) in StatusBarManager (publishers already @MainActor)
- Remove redundant #available(macOS 13) guards in LaunchAtLoginSection (deployment target is macOS 13)
- Update spec/ARCHITECTURE.md project tree and spec/CONSTANTS.md
