+++ AIBattery/Views/UsagePopoverView.swift
@@ -602,308 +602,3 @@ private struct ActivityChartGate: View {
         }
     }
 }
-
-// MARK: - Settings (decomposed into focused sub-views)
-
-
-/// Inline settings for account names, refresh rate, and notifications.
-private struct SettingsRow: View {
-    let viewModel: UsageViewModel
-    @ObservedObject var accountStore: AccountStore
-    let onAddAccount: () -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            Text("Settings")
-                .font(.caption.weight(.semibold))
-                .foregroundStyle(.secondary)
-
-            // Per-account names
-            ForEach(Array(accountStore.accounts.enumerated()), id: \.element.id) { index, account in
-                accountNameRow(account, index: index)
-            }
-
-            if accountStore.canAddAccount {
-                HStack(spacing: 8) {
-                    Spacer().frame(width: 50)
-                    Button(action: onAddAccount) {
-                        HStack(spacing: 3) {
-                            Image(systemName: "plus.circle")
-                                .font(.system(size: 10))
-                            Text("Add Account")
-                                .font(.caption)
-                        }
-                    }
-                    .buttonStyle(.plain)
-                    .foregroundStyle(.blue)
-                    .accessibilityLabel("Add another Claude account")
-                }
-            }
-
-            RefreshSettingsSection(viewModel: viewModel)
-            DisplaySettingsSection()
-            AlertSettingsSection()
-            LaunchAtLoginSection()
-        }
-        .padding(.horizontal, 16)
-        .padding(.vertical, 10)
-    }
-
-    /// Editable name row for a single account.
-    private func accountNameRow(_ account: AccountRecord, index: Int) -> some View {
-        let isActive = account.id == accountStore.activeAccountId
-        let label = accountStore.accounts.count > 1
-            ? (isActive ? "Active" : "Account")
-            : "Name"
-        return HStack(spacing: 8) {
-            Text(label)
-                .font(.caption)
-                .foregroundStyle(.secondary)
-                .frame(width: 50, alignment: .trailing)
-            TextField("Account \(index + 1)", text: nameBinding(for: account.id))
-                .textFieldStyle(.roundedBorder)
-                .font(.caption)
-            if accountStore.accounts.count > 1 {
-                Button(action: {
-                    OAuthManager.shared.signOut(accountId: account.id)
-                }) {
-                    Image(systemName: "xmark.circle")
-                        .font(.system(size: 10))
-                        .foregroundStyle(.secondary)
-                }
-                .buttonStyle(.plain)
-                .help("Remove this account")
-                .accessibilityLabel("Remove account \(index + 1)")
-            }
-        }
-    }
-
-    /// Two-way binding that reads/writes `displayName` on the AccountRecord.
-    private func nameBinding(for accountId: String) -> Binding<String> {
-        Binding(
-            get: {
-                accountStore.accounts.first { $0.id == accountId }?.displayName ?? ""
-            },
-            set: { newValue in
-                let clamped = String(newValue.prefix(30))
-                OAuthManager.shared.updateAccountMetadata(
-                    accountId: accountId,
-                    displayName: clamped
-                )
-            }
-        )
-    }
-}
-
-/// Refresh interval slider.
-private struct RefreshSettingsSection: View {
-    let viewModel: UsageViewModel
-    @AppStorage(UserDefaultsKeys.refreshInterval) private var refreshInterval: Double = 60
-
-    var body: some View {
-        VStack(spacing: 2) {
-            HStack(spacing: 8) {
-                Text("Refresh")
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-                    .frame(width: 50, alignment: .trailing)
-                Slider(value: $refreshInterval, in: 10...60, step: 5)
-                    .onChange(of: refreshInterval) { _ in
-                        viewModel.updatePollingInterval(refreshInterval)
-                    }
-                    .accessibilityLabel("Refresh interval")
-                    .accessibilityValue("\(Int(refreshInterval)) seconds")
-                Text("\(Int(refreshInterval))s")
-                    .font(.system(.caption, design: .monospaced))
-                    .frame(width: 28, alignment: .trailing)
-            }
-            sliderMarks(labels: ["10s", "20s", "30s", "40s", "50s", "60s"], leadingPad: 50)
-            Text("~3 tokens per poll")
-                .font(.caption2)
-                .foregroundStyle(ThemeColors.tertiaryLabel)
-                .padding(.leading, 58)
-        }
-    }
-}
-
-/// Display toggles + idle session cutoff slider.
-private struct DisplaySettingsSection: View {
-    @AppStorage(UserDefaultsKeys.idleSessionMinutes) private var idleSessionMinutes: Double = 0
-    @AppStorage(UserDefaultsKeys.showTokens) private var showTokens: Bool = true
-    @AppStorage(UserDefaultsKeys.showActivity) private var showActivity: Bool = true
-    @AppStorage(UserDefaultsKeys.colorblindMode) private var colorblindMode: Bool = false
-    @AppStorage(UserDefaultsKeys.showCostEstimate) private var showCostEstimate: Bool = false
-
-    /// Slider positions (1–6) mapped to minutes: 30, 60, 120, 240, 480, 0 (never).
-    private static let idleSteps: [Double] = [30, 60, 120, 240, 480, 0]
-
-    var body: some View {
-        // Idle session cutoff (slider 1–6; positions map to 30m/1h/2h/4h/8h/Never)
-        VStack(spacing: 2) {
-            HStack(spacing: 8) {
-                Text("Hide idle")
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-                    .frame(width: 50, alignment: .trailing)
-                Slider(value: idleSliderBinding, in: 1...6, step: 1)
-                    .accessibilityLabel("Hide idle sessions")
-                    .accessibilityValue(idleLabel)
-                Text(idleLabel)
-                    .font(.system(.caption, design: .monospaced))
-                    .frame(width: 28, alignment: .trailing)
-            }
-            sliderMarks(labels: ["30m", "1h", "2h", "4h", "8h", "\u{221E}"], leadingPad: 50)
-        }
-
-        // Display toggles
-        HStack(spacing: 8) {
-            Text("Display")
-                .font(.caption)
-                .foregroundStyle(.secondary)
-                .frame(width: 50, alignment: .trailing)
-            Toggle("Tokens", isOn: $showTokens)
-                .toggleStyle(.checkbox)
-                .font(.caption)
-            Toggle("Activity", isOn: $showActivity)
-                .toggleStyle(.checkbox)
-                .font(.caption)
-        }
-        HStack(spacing: 8) {
-            Spacer().frame(width: 50)
-            Toggle("Colorblind", isOn: $colorblindMode)
-                .toggleStyle(.checkbox)
-                .font(.caption)
-                .help("Use colorblind-safe palette (blue/cyan/amber/purple)")
-            Toggle("Cost*", isOn: $showCostEstimate)
-                .toggleStyle(.checkbox)
-                .font(.caption)
-        }
-        HStack(spacing: 8) {
-            Spacer().frame(width: 50)
-            Text("Cost* = equivalent API token rates")
-                .font(.caption2)
-                .foregroundStyle(ThemeColors.tertiaryLabel)
-        }
-    }
-
-    /// Maps slider position (1–6) ↔ stored minutes (30/60/120/240/480/0).
-    private var idleSliderBinding: Binding<Double> {
-        Binding(
-            get: {
-                if let idx = Self.idleSteps.firstIndex(of: idleSessionMinutes) {
-                    return Double(idx + 1)
-                }
-                return 6 // default to "Never"
-            },
-            set: { idleSessionMinutes = Self.idleSteps[max(0, min(Int($0) - 1, Self.idleSteps.count - 1))] }
-        )
-    }
-
-    /// Display label for the current idle cutoff.
-    private var idleLabel: String {
-        switch Int(idleSessionMinutes) {
-        case 30: return "30m"
-        case 60: return "1h"
-        case 120: return "2h"
-        case 240: return "4h"
-        case 480: return "8h"
-        default: return "\u{221E}"
-        }
-    }
-}
-
-/// Status alerts + rate limit alerts.
-private struct AlertSettingsSection: View {
-    @AppStorage(UserDefaultsKeys.alertStatus) private var alertStatus: Bool = false
-    @AppStorage(UserDefaultsKeys.alertRateLimit) private var alertRateLimit: Bool = false
-    @AppStorage(UserDefaultsKeys.rateLimitThreshold) private var rateLimitThreshold: Double = 80
-
-    var body: some View {
-        HStack(spacing: 8) {
-            Text("Alerts")
-                .font(.caption)
-                .foregroundStyle(.secondary)
-                .frame(width: 50, alignment: .trailing)
-            Toggle("Status", isOn: $alertStatus)
-                .toggleStyle(.checkbox)
-                .font(.caption)
-                .onChange(of: alertStatus) { on in
-                    if on { NotificationManager.shared.requestPermission() }
-                }
-            Toggle("Rate Limit", isOn: $alertRateLimit)
-                .toggleStyle(.checkbox)
-                .font(.caption)
-                .onChange(of: alertRateLimit) { on in
-                    if on { NotificationManager.shared.requestPermission() }
-                }
-            if alertStatus {
-                Button("Test") {
-                    NotificationManager.shared.testAlerts()
-                }
-                .buttonStyle(.plain)
-                .font(.caption2)
-                .foregroundStyle(.blue)
-            }
-        }
-        if alertRateLimit {
-            VStack(spacing: 2) {
-                HStack(spacing: 8) {
-                    Spacer().frame(width: 50)
-                    Slider(value: $rateLimitThreshold, in: 50...95, step: 5)
-                        .accessibilityLabel("Rate limit alert threshold")
-                        .accessibilityValue("\(Int(rateLimitThreshold)) percent")
-                    Text("\(Int(rateLimitThreshold))%")
-                        .font(.system(.caption, design: .monospaced))
-                        .frame(width: 28, alignment: .trailing)
-                }
-                sliderMarks(labels: ["50%", "60%", "70%", "80%", "90%", "95%"], leadingPad: 50)
-            }
-        }
-    }
-}
-
-/// Launch at Login toggle.
-private struct LaunchAtLoginSection: View {
-    @AppStorage(UserDefaultsKeys.launchAtLogin) private var launchAtLogin: Bool = false
-
-    var body: some View {
-        HStack(spacing: 8) {
-            Text("Startup")
-                .font(.caption)
-                .foregroundStyle(.secondary)
-                .frame(width: 50, alignment: .trailing)
-            Toggle("Launch at Login", isOn: $launchAtLogin)
-                .toggleStyle(.checkbox)
-                .font(.caption)
-                .onAppear {
-                    if #available(macOS 13.0, *) {
-                        launchAtLogin = LaunchAtLoginManager.isEnabled
-                    }
-                }
-                .onChange(of: launchAtLogin) { newValue in
-                    if #available(macOS 13.0, *) {
-                        LaunchAtLoginManager.setEnabled(newValue)
-                    }
-                }
-        }
-    }
-}
-
-/// Tick mark labels displayed below a slider. Shared across settings sections.
-fileprivate func sliderMarks(labels: [String], leadingPad: CGFloat) -> some View {
-    HStack {
-        Spacer().frame(width: leadingPad + 8) // label width + HStack spacing
-        HStack(spacing: 0) {
-            ForEach(Array(labels.enumerated()), id: \.offset) { i, label in
-                Text(label)
-                    .font(.system(size: 8))
-                    .foregroundStyle(ThemeColors.tertiaryLabel)
-                if i < labels.count - 1 {
-                    Spacer()
-                }
-            }
-        }
-        Spacer().frame(width: 36) // value label width + spacing
-    }
-}
-
