+++ spec/CONSTANTS.md
@@ -342,7 +342,7 @@ Most bar and accent colors use system palette in both modes (the opaque light-mo
 |------|---------|--------|
 | `ENABLE_SPARKLE` | Defined (all 3 SPM targets) | Guards Sparkle imports, `SparkleUpdateService`, and Sparkle-dependent UI. Remove to build App Store variant without Sparkle. |
 | `ENABLE_VERSION_CHECKER` | Defined (all 3 SPM targets) | Guards `VersionChecker`, update banner UI, and version check button. Remove for App Store (guideline 3.1.1). |
-| `APP_SANDBOX` | NOT defined | Enables `SandboxAccessManager` (security-scoped bookmark for `~/.claude/` access). Only set for App Store builds. |
+| `APP_SANDBOX` | NOT defined | Reserved for future App Store sandbox support (security-scoped bookmark for `~/.claude/` access). Only set for App Store builds. |
 
 ## Build Environment Variables
